+++ spec/store_model/model_spec.rb
@@ -59,13 +59,39 @@ RSpec.describe StoreModel::Model do
     context "when two instances have same attributes" do
       let(:second_setting) { Configuration.new(color: "red") }
 
-      it { is_expected.to be_truthy }
+      it { is_expected.to be true }
     end
 
     context "when two instances have different attributes" do
-      let(:second_setting) { Configuration.new(color: "user") }
+      let(:second_setting) { Configuration.new(color: "black") }
 
-      it { is_expected.to be_falsey }
+      it { is_expected.to be false }
+    end
+
+    context "when StoreModel has enum attribute" do
+      let(:config_class) do
+        Class.new do
+          include StoreModel::Model
+
+          enum :status, in: { active: 1, archived: 0 }
+        end
+      end
+
+      let(:first_setting) { config_class.new(status: :active) }
+
+      subject { first_setting == second_setting }
+
+      context "when two instances have same attributes" do
+        let(:second_setting) { config_class.new(status: :active) }
+
+        it { is_expected.to be true }
+      end
+
+      context "when two instances have different attributes" do
+        let(:second_setting) { config_class.new(status: :archived) }
+
+        it { is_expected.to be false }
+      end
     end
   end